Make some Sanity queries reusable, for previews
